Remove markdown parser.
I tried two parsers, both function fine but contrast sharply with the
exercise I'm following. There is no need to create a Post struct from
the files if using a package that parses the markdown.

So, for the sake of continuing with the exercise, I'm not going to
convert the Body to HTML.